fix if

+++ vars/cleanConfig.groovy
@@ -12,7 +12,7 @@ def call(project) {
 	}
 	
 	// Clean deployment configurations
-	f((project == devProject && dcDevChanged.toBoolean()) || (project == testProject && dcTestChanged.toBoolean())){
+	if((project == devProject && dcDevChanged.toBoolean()) || (project == testProject && dcTestChanged.toBoolean())){
 		command = "oc process -f cicd/iamp-spring-service-template.yaml -p APP_NAME=${appName} -p IMAGE_NAME=${appName} -p IMAGE_TAG=latest | oc delete -f- -n " + project + " || true"
 		
 		sh command
